+++ src/ResourceManager/Sql/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Current Release
 * Updated example in the help file for Get-AzureRmSqlDatabaseExpanded
+* Updated Set-AzureRmSqlServerAuditing and Set-AzureRmSqlDatabaseAuditing Cmdlets
+    - Added a new parameter called StorageAccountSubscriptionId.
+	- Added a new example for each modified cmdlet.
 
 ## Version 4.6.0
 * Updated the following cmdlets with optional LicenseType parameter